Updates to #31
I just verified that the applied fix worked and I also cleaned up a unit
test that is a todo.. We should get that scenario working (the commented
out test).

@ejsmith I don't see any problems with this pull request. I think we
just need todo some testing.. Do you remember why we weren't doing this
before (setting the contract resolver) or was this just an oversight?