Fixed an issue with self hosted installs that used AppSettings.
The scheme would be set to 443 always when it shouldn't of been set.
This code is now driven off of the scheme specified. @ejsmith  I don't
even think we need this code block anymore because the underlying uri
parser looks to be setting the port correctly based on the scheme.