Added flag to see if the configuration is currently updating
Config header checks can force an update as well as the settings timer running on the client at the same time.

One problem with this is that SettingsManager is stateless and could have mutliple clients using it. This change can prevent multiple clients from updating at the same time. There are already checks in the client instance but they are not visible to the submission client